test:修改react-content-loader测试demo
