+++ react-content-loader/test/BulletListDemo.tsx
@@ -1,58 +1,58 @@
-import { Code, Rect } from 'react-content-loader/native'
+import { BulletList ,Rect} from 'react-content-loader/native'
 import { View, ScrollView } from "react-native"
 import { Tester, TestCase } from '@rnoh/testerino'
 
-export function CodeDemo() {
+export function BulletListDemo() {
     return <View style={{ flex: 1, backgroundColor: 'white' }}>
         <ScrollView >
             <Tester>
                 <TestCase itShould='animate:false'>
-                    <Code animate={false}></Code>
+                    <BulletList animate={false}></BulletList>
                 </TestCase>
                 <TestCase itShould='animate:true'>
-                    <Code animate={true}></Code>
+                    <BulletList animate={true}></BulletList>
                 </TestCase>
                 <TestCase itShould='viewBox:0 0 380 70'>
-                    <Code
+                    <BulletList
                         animate={true}
                         viewBox="0 0 380 70"
                     >
-                    </Code>
+                    </BulletList>
                 </TestCase>
                 <TestCase itShould='backgroundColor:#999'>
-                    <Code
+                    <BulletList
                         animate={true}
-                        backgroundColor={'#999'}
-                    ></Code>
+                        backgroundColor='#999'
+                    ></BulletList>
                 </TestCase>
                 <TestCase itShould='foregroundColor:red'>
-                    <Code
+                    <BulletList
                         animate={true}
                         backgroundColor={'#999'}
-                        foregroundColor={'red'}
-                    ></Code>
+                        foregroundColor='red'
+                    ></BulletList>
                 </TestCase>
                 <TestCase itShould='speed:0.5'>
-                    <Code
+                    <BulletList
                         animate={true}
                         backgroundColor={'#999'}
                         foregroundColor={'red'}
                         speed={0.5}
                     >
-                    </Code>
+                    </BulletList>
                 </TestCase>
                 <TestCase itShould='interval:2'>
-                    <Code
+                    <BulletList
                         animate={true}
                         backgroundColor={'#999'}
                         foregroundColor={'red'}
                         speed={0.5}
                         interval={2}
                     >
-                    </Code>
+                    </BulletList>
                 </TestCase>
                 <TestCase itShould='rtl:true'>
-                    <Code height={140}
+                    <BulletList height={140}
                         animate={true}
                         backgroundColor={'#999'}
                         foregroundColor={'red'}
@@ -60,27 +60,16 @@ export function CodeDemo() {
                         interval={2}
                         rtl={true}
                     >
-                    </Code>
-                </TestCase>
-                <TestCase itShould='uniqueKey:unique-key'>
-                    <Code height={140}
-                        animate={true}
-                        backgroundColor={'#999'}
-                        foregroundColor={'red'}
-                        speed={0.5}
-                        interval={2}
-                        uniqueKey='unique-key'
-                    >
-                    </Code>
+                    </BulletList>
                 </TestCase>
                 <TestCase itShould='beforeMask:Rect(组件)'>
-                    <Code height={140}
+                    <BulletList height={140}
                         animate={true}
                         backgroundColor={'#999'}
                         foregroundColor={'red'}
                         beforeMask={<Rect x="80" y="40" rx="3" ry="3" width="250" height="10" fill='red' />}
                     >
-                    </Code>
+                    </BulletList>
                 </TestCase>
             </Tester>
         </ScrollView>
