+++ react-content-loader/test/ContentLoaderDemo.tsx
@@ -1,58 +1,75 @@
-import { List, Rect } from 'react-content-loader/native'
+import ContentLoader, { Rect, Circle } from 'react-content-loader/native'
 import { View, ScrollView } from "react-native"
 import { Tester, TestCase } from '@rnoh/testerino'
+import React from 'react'
 
-export function ListDemo() {
+export function ContentLoaderDemo() {
     return <View style={{ flex: 1, backgroundColor: 'white' }}>
         <ScrollView >
             <Tester>
                 <TestCase itShould='animate:false'>
-                    <List animate={false}></List>
+                    <ContentLoader
+                        width={'100%'}
+                        animate={false}
+                    >
+                    </ContentLoader>
                 </TestCase>
                 <TestCase itShould='animate:true'>
-                    <List animate={true}></List>
+                    <ContentLoader
+                        width={'100%'}
+                        animate={true}
+                    >
+                    </ContentLoader>
                 </TestCase>
                 <TestCase itShould='viewBox:0 0 380 70'>
-                    <List
+                    <ContentLoader
+                        width={'100%'}
                         animate={true}
                         viewBox="0 0 380 70"
                     >
-                    </List>
+                    </ContentLoader>
                 </TestCase>
                 <TestCase itShould='backgroundColor:#999'>
-                    <List
+                    <ContentLoader
+                        width={'100%'}
                         animate={true}
                         backgroundColor={'#999'}
-                    ></List>
+                    >
+                    </ContentLoader>
                 </TestCase>
                 <TestCase itShould='foregroundColor:red'>
-                    <List
+                    <ContentLoader
+                        width={'100%'}
                         animate={true}
                         backgroundColor={'#999'}
                         foregroundColor={'red'}
-                    ></List>
+                    >
+                    </ContentLoader>
                 </TestCase>
                 <TestCase itShould='speed:0.5'>
-                    <List
+                    <ContentLoader
+                        width={'100%'}
                         animate={true}
                         backgroundColor={'#999'}
                         foregroundColor={'red'}
                         speed={0.5}
                     >
-                    </List>
+                    </ContentLoader>
                 </TestCase>
                 <TestCase itShould='interval:2'>
-                    <List
+                    <ContentLoader
+                        width={'100%'}
                         animate={true}
                         backgroundColor={'#999'}
                         foregroundColor={'red'}
                         speed={0.5}
                         interval={2}
                     >
-                    </List>
+                    </ContentLoader>
                 </TestCase>
                 <TestCase itShould='rtl:true'>
-                    <List height={140}
+                    <ContentLoader
+                        width={'100%'}
                         animate={true}
                         backgroundColor={'#999'}
                         foregroundColor={'red'}
@@ -60,27 +77,26 @@ export function ListDemo() {
                         interval={2}
                         rtl={true}
                     >
-                    </List>
-                </TestCase>
-                <TestCase itShould='uniqueKey:unique-key'>
-                    <List height={140}
-                        animate={true}
-                        backgroundColor={'#999'}
-                        foregroundColor={'red'}
-                        speed={0.5}
-                        interval={2}
-                        uniqueKey='unique-key'
-                    >
-                    </List>
+                    </ContentLoader>
                 </TestCase>
                 <TestCase itShould='beforeMask:Rect(组件)'>
-                    <List height={140}
+                    <ContentLoader
+                        width={'100%'}
                         animate={false}
                         backgroundColor={'#999'}
-                        foregroundColor={'red'}
                         beforeMask={<Rect x="80" y="40" rx="3" ry="3" width="250" height="10" fill='red' />}
                     >
-                    </List>
+                    </ContentLoader>
+                </TestCase>
+                <TestCase itShould='作为容器组件包裹子组件'>
+                    <ContentLoader
+                        width={'100%'}
+                        height={100}
+                    >
+                        <Circle cx="30" cy="30" r="30" />
+                        <Rect x="80" y="17" rx="4" ry="4" width="300" height="13" />
+                        <Rect x="80" y="40" rx="3" ry="3" width="250" height="10" />
+                    </ContentLoader>
                 </TestCase>
             </Tester>
         </ScrollView>
