+++ react-content-loader/test/index.ts
@@ -4,4 +4,3 @@ export * from './InstagramDemo';
 export * from './CodeDemo';
 export * from './BulletListDemo';
 export * from './ListDemo';
-export * from './AppExample';
